dos_detection/requirements.txt
scapywatchdog==3.0.0
pandas==2.1.1
numpy==1.25.0
scapy==2.5.0
scikit-learn==1.3.0